requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==3.0.0
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
whitenoise==6.9.0
